libm: Sync with NetBSD
Adds the following assembly function: logbl, ilogbl, modf, scalbnl
Adds the following source functions: exp2, exp2f, frexp, nextafterl,
  nexttoward, requo, remquof, and fabsl
Captures all the bug fixes since last sync
Synchronizes with current NetBSD versions where possible.

Required to fix modern gcc/fortran and other packages.
Gets closer to C99 capability.

+++ lib/libm/src/e_rem_pio2f.c
@@ -12,8 +12,7 @@
  * is preserved.
  * ====================================================
  *
- * $NetBSD: e_rem_pio2f.c,v 1.8 2002/05/26 22:01:52 wiz Exp $
- * $DragonFly: src/lib/libm/src/e_rem_pio2f.c,v 1.1 2005/07/26 21:15:20 joerg Exp $
+ * $NetBSD: e_rem_pio2f.c,v 1.9 2009/01/19 06:00:30 lukem Exp $
  */
 
 /* __libm_rem_pio2f(x,y)
@@ -128,7 +127,7 @@ __libm_rem_pio2f(float x, float *y)
 	    fn = (float)n;
 	    r  = t-fn*pio2_1;
 	    w  = fn*pio2_1t;	/* 1st round good to 40 bit */
-	    if(n<32&&(ix&0xffffff00)!=npio2_hw[n-1]) {
+	    if(n<32&&(int32_t)(ix&0xffffff00)!=npio2_hw[n-1]) {
 		y[0] = r-w;	/* quick check no cancellation */
 	    } else {
 	        u_int32_t high;
