+++ lib/libm/src/lroundf.c
@@ -1,3 +1,5 @@
+/* $NetBSD: lroundf.c,v 1.3 2008/04/26 23:49:50 christos Exp $ */
+
 /*-
  * Copyright (c) 2004
  *	Matthias Drochner. All rights reserved.
@@ -22,9 +24,6 @@
  * LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY
  * OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF
  * SUCH DAMAGE.
- *
- * $NetBSD: lroundf.c,v 1.2 2004/10/13 15:18:32 drochner Exp $
- * $DragonFly: src/lib/libm/src/lroundf.c,v 1.1 2005/07/26 21:15:20 joerg Exp $
  */
 
 #include <math.h>
@@ -50,7 +49,7 @@ LROUNDNAME(float x)
 
 	GET_FLOAT_WORD(i0, x);
 	e = i0 >> SNG_FRACBITS;
-	s = e >> SNG_EXPBITS;
+	s = (uint32_t)e >> SNG_EXPBITS;
 	e = (e & 0xff) - SNG_EXP_BIAS;
 
 	/* 1.0 x 2^-1 is the smallest number which can be rounded to 1 */
