+++ lib/libm/src/s_scalbn.c
@@ -9,8 +9,7 @@
  * is preserved.
  * ====================================================
  *
- * $NetBSD: s_scalbn.c,v 1.12 2002/05/26 22:01:58 wiz Exp $
- * $DragonFly: src/lib/libm/src/s_scalbn.c,v 1.1 2005/07/26 21:15:20 joerg Exp $
+ * $NetBSD: s_scalbn.c,v 1.15 2011/07/26 16:10:16 joerg Exp $
  */
 
 /*
@@ -34,12 +33,12 @@ scalbn(double x, int n)
 {
 	int32_t k,hx,lx;
 	EXTRACT_WORDS(hx,lx,x);
-        k = (hx&0x7ff00000)>>20;		/* extract exponent */
+        k = ((uint32_t)hx&0x7ff00000)>>20;		/* extract exponent */
         if (k==0) {				/* 0 or subnormal x */
             if ((lx|(hx&0x7fffffff))==0) return x; /* +-0 */
 	    x *= two54;
 	    GET_HIGH_WORD(hx,x);
-	    k = ((hx&0x7ff00000)>>20) - 54;
+	    k = (((uint32_t)hx&0x7ff00000)>>20) - 54;
             if (n< -50000) return tiny*x; 	/*underflow*/
 	    }
         if (k==0x7ff) return x+x;		/* NaN or Inf */
